Moved classes to folders
