+++ app.rb
@@ -1,12 +1,12 @@
 require 'json'
-require_relative 'game'
-require_relative 'author'
+require_relative './classes/game'
+require_relative './classes/author'
 require_relative 'data_module'
-require_relative 'music_album'
-require_relative 'genre'
-require_relative 'book'
+require_relative './classes/music_album'
+require_relative './classes/genre'
+require_relative './classes/book'
 require_relative 'storage'
-require_relative 'label'
+require_relative './classes/label'
 
 # class application
 class Application
