+++ data/albums.json
@@ -1,1 +1,1 @@
-[{"publish_date":"2023-05-30","on_spotify":true}]
+[{"publish_date":"2023-05-30","on_spotify":true},{"publish_date":"2023-05-31","on_spotify":true}]
