+++ data/author_list.json
@@ -14,5 +14,9 @@
   {
     "first_name": "Jose",
     "last_name": "Test"
+  },
+  {
+    "first_name": "Mehmed",
+    "last_name": "Second"
   }
 ]
