--- data/authors.json
@@ -1,27 +1,0 @@
-[
-  {
-    "id": 1,
-    "first_name": "kkkk",
-    "last_name": "lllll"
-  },
-  {
-    "id": 2,
-    "first_name": "first",
-    "last_name": "last"
-  },
-  {
-    "id": 3,
-    "first_name": "Kola",
-    "last_name": "Kolade"
-  },
-  {
-    "id": 4,
-    "first_name": "Banjo",
-    "last_name": "Big boy"
-  },
-  {
-    "id": 5,
-    "first_name": "Bruce",
-    "last_name": "Lee"
-  }
-]
