+++ data/game_list.json
@@ -3,5 +3,10 @@
     "id": null,
     "last_played_at": "2023-05-30",
     "multiplayer": false
+  },
+  {
+    "id": 930,
+    "last_played_at": "2023-03-03",
+    "multiplayer": true
   }
 ]
