+++ data/genres.json
@@ -1,1 +1,1 @@
-[{"id":222,"name":"Comedy"},{"id":132,"name":"Best"},{"id":424,"name":"Rock"},{"id":220,"name":"Rock"},{"id":510,"name":"Terror"}]
+[{"id":222,"name":"Comedy"},{"id":132,"name":"Best"},{"id":424,"name":"Rock"},{"id":220,"name":"Rock"},{"id":510,"name":"Terror"},{"id":670,"name":"Documental"}]
