--- data/sources.json
@@ -1,22 +1,0 @@
-[
-  {
-    "id": 1,
-    "name": "gift"
-  },
-  {
-    "id": 2,
-    "name": "purchase"
-  },
-  {
-    "id": 3,
-    "name": "Borrowed"
-  },
-  {
-    "id": 4,
-    "name": "Loan"
-  },
-  {
-    "id": 5,
-    "name": "Friend"
-  }
-]
